Pull model into separate file
